Split Software.Henforcer into submodules
